set the new grid in the list component, I will finish the styling on sunday
