+++ src/components/sale/card.js
@@ -2,38 +2,11 @@ import React, {useState} from "react";
 import { Link } from "react-router-dom";
 import Grid from '@material-ui/core/Grid';
 import './card.css'
-import Table from '@material-ui/core/Table';
-import { makeStyles, withStyles } from '@material-ui/core/styles';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TableRow from '@material-ui/core/TableRow';
-import Paper from '@material-ui/core/Paper';
-
-
 import Modal from 'react-bootstrap/Modal';
 
 const SaleCard = (props) => {
 
-  const StyledTableCell = withStyles((theme) => ({
-    head: {
-      backgroundColor: theme.palette.common.black,
-      color: theme.palette.common.white,
-    },
-    body: {
-      fontSize: 14,
-    },
-  }))(TableCell);
-  
-  const StyledTableRow = withStyles((theme) => ({
-    root: {
-      '&:nth-of-type(odd)': {
-        backgroundColor: theme.palette.action.hover,
-      },
-    },
-
-  }))(TableRow);  
+ 
 
     const [show, setShow] = useState(false);
 
@@ -44,32 +17,7 @@ const SaleCard = (props) => {
         <>
           <div onClick={() => handleShow()} className="sale-card--container">
           <div className={props.classes.root}>
-          <TableContainer component={Paper}>
-      <Table className={props.classes.table} aria-label="customized table">
-        <TableHead>
-          <TableRow>
-          <StyledTableCell>#invoice number</StyledTableCell>
-            <StyledTableCell align="right">purchase_date</StyledTableCell>
-            <StyledTableCell align="right">vehicle</StyledTableCell>
-            <StyledTableCell align="right">sale_type</StyledTableCell>
-            <StyledTableCell align="right">$sale_price</StyledTableCell>
-          </TableRow>
-        </TableHead>
-        <TableBody>
-          {props.item.map((row) => (
-            <StyledTableRow key={row.name}>
-              <StyledTableCell component="th" scope="row">
-                {row.name}
-              </StyledTableCell>
-              <StyledTableCell align="right">{row.calories}</StyledTableCell>
-              <StyledTableCell align="right">{row.fat}</StyledTableCell>
-              <StyledTableCell align="right">{row.carbs}</StyledTableCell>
-              <StyledTableCell align="right">{row.protein}</StyledTableCell>
-            </StyledTableRow>
-          ))}
-        </TableBody>
-      </Table>
-    </TableContainer>
+         
           </div>
           </div>
